Spawn trees

+++ Scripts/InventorySystem/ToolData.cs
@@ -7,6 +7,6 @@ namespace Scripts.InventorySystem;
 public partial class ToolData : ItemData
 {
     [Export] public float Damage = 0.0f;
-    [Export] public float GatheringTime = 0.0f;
+    [Export] public double GatheringTime = 0.0f;
     [Export] public bool Equippable = true;
 }
